added seperate files for test utils and test_usj functions
