+++ rust-usfm-parser/src/input.usfm
@@ -1,0 +1,3 @@
+\id GEN
+\c 1
+\p \v 1 \em word\em* \em wordtwo\em* word3 \em word4 \em* \v 2 \em word5 \em* \v 3 \w glossaryone\w* \w glossarytwo\w*r
